+++ AlphaFS/Filesystem/File Class/File.Create.cs
@@ -363,7 +363,7 @@ namespace Alphaleonis.Win32.Filesystem
          // the path string should be the following form: "\\.\X:"
          // Do not use a trailing backslash (\), which indicates the root.
 
-         string pathLp = Path.GetExtendedLengthPathCore(transaction, path, pathFormat, GetFullPathOptions.TrimEnd | GetFullPathOptions.RemoveTrailingDirectorySeparator);
+         var pathLp = Path.GetExtendedLengthPathCore(transaction, path, pathFormat, GetFullPathOptions.TrimEnd | GetFullPathOptions.RemoveTrailingDirectorySeparator);
 
          PrivilegeEnabler privilegeEnabler = null;
 
@@ -393,7 +393,7 @@ namespace Alphaleonis.Win32.Filesystem
          using (privilegeEnabler)
          using (var securityAttributes = new Security.NativeMethods.SecurityAttributes(fileSecurity))
          {
-            SafeFileHandle handle = transaction == null || !NativeMethods.IsAtLeastWindowsVista
+            var handle = transaction == null || !NativeMethods.IsAtLeastWindowsVista
 
                // CreateFile() / CreateFileTransacted()
                // In the ANSI version of this function, the name is limited to MAX_PATH characters.
@@ -403,7 +403,7 @@ namespace Alphaleonis.Win32.Filesystem
                ? NativeMethods.CreateFile(pathLp, fileSystemRights, fileShare, securityAttributes, fileMode, attributes, IntPtr.Zero)
                : NativeMethods.CreateFileTransacted(pathLp, fileSystemRights, fileShare, securityAttributes, fileMode, attributes, IntPtr.Zero, transaction.SafeHandle, IntPtr.Zero, IntPtr.Zero);
 
-            int lastError = Marshal.GetLastWin32Error();
+            var lastError = Marshal.GetLastWin32Error();
 
             if (handle.IsInvalid)
             {
@@ -414,7 +414,7 @@ namespace Alphaleonis.Win32.Filesystem
 
             if (isAppend)
             {
-               var stream = new FileStream(handle, FileAccess.Write, 4096, (attributes & ExtendedFileAttributes.Overlapped) != 0);
+               var stream = new FileStream(handle, FileAccess.Write, NativeMethods.DefaultFileBufferSize, (attributes & ExtendedFileAttributes.Overlapped) != 0);
                stream.Seek(0, SeekOrigin.End);
             }